refactor: 패키지 구조 변경
- 앱 시작 시 데이터 초기화 코드 이동
- init 패키지 추가
